+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.0
+
+- Bumped Android SDK to `3.1.3`.
+- Bumped iOS SDK to `1.2.0`.
+- Removed `allowMultipleDocuments`.
+- Removed `defaultMultipleDocuments`.
+- Added `DocumentMode`.
+- Added `cameraModeSingle`, `cameraModeMulti`, `cameraModeSegmented` which are all of type `DocumentMode`.
+- Fixed typo `imageColorGrayscaleText` becomes `imageColorGrayScaleText`.
+
 ## 0.0.2
 
 - Fixed typo`Filepath` becomes `filePath` in Android SDK.